Switch to app router edge runtime
